fix: Fixed a bug related to dependencies check.
The pass was not able to correctly identify negative distance
dependencies. Now it can do it by leveraging ScalarEvolution analysis.
Specifically it get the add recurrence of the two dependent instructions
and computes the distance. If the distance is a compile-time constant
and it is negative, then there is a negative distance dependency and the
loop fusion cannot be applied.