fix: fixed a bug related to nested loop fusion.
When trying to fuse nested loops the pass was not able to properly fuse
the nests. Now, it is able to fuse loop nests greedily, obtaining, in
some cases, even a single loop nest.

e.g.

Before the fix the pass fused only the two outer loops. Now it fuses
even the two inner loops.

for (int i = 0; i < 3; i++) {
	for (int j = 0; j < 3; j++) {
	    a[i][j] = a[i][j] + 1;
	}
}

for (int i = 0; i < 3; i++) {
	for (int j = 0; j < 3; j++) {
	    b[i][j] = a[i][j] + 1;
}

It is important to note that It won't fuse any of the loops if there is
a dependency either on the i or the j index.